+++ package/abis/LaborMarket.ts
@@ -28,6 +28,12 @@ export const abi = [
       },
       {
         "indexed": false,
+        "internalType": "string",
+        "name": "uri",
+        "type": "string"
+      },
+      {
+        "indexed": false,
         "internalType": "address",
         "name": "criteria",
         "type": "address"
@@ -478,6 +484,11 @@ export const abi = [
         "internalType": "address",
         "name": "_deployer",
         "type": "address"
+      },
+      {
+        "internalType": "string",
+        "name": "_uri",
+        "type": "string"
       },
       {
         "internalType": "contract EnforcementCriteriaInterface",