+++ package/abis/LaborMarketFactoryInterface.ts
@@ -32,6 +32,11 @@ export const abi = [
         "type": "address"
       },
       {
+        "internalType": "string",
+        "name": "_uri",
+        "type": "string"
+      },
+      {
         "internalType": "contract EnforcementCriteriaInterface",
         "name": "_criteria",
         "type": "address"