+++ package/types/factories/src/interfaces/LaborMarketInterface__factory.ts
@@ -21,6 +21,12 @@ const _abi = [
       },
       {
         indexed: false,
+        internalType: "string",
+        name: "uri",
+        type: "string",
+      },
+      {
+        indexed: false,
         internalType: "address",
         name: "criteria",
         type: "address",
@@ -376,6 +382,11 @@ const _abi = [
         internalType: "address",
         name: "_deployer",
         type: "address",
+      },
+      {
+        internalType: "string",
+        name: "_uri",
+        type: "string",
       },
       {
         internalType: "contract EnforcementCriteriaInterface",